// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ROLE = {
  CUSTOMER: 'CUSTOMER',
  SELLER: 'SELLER',
  ADMIN: 'ADMIN'
} as const

export type ROLE = (typeof ROLE)[keyof typeof ROLE]


export const ORDER_STATUS = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
  PLACED: 'PLACED'
} as const

export type ORDER_STATUS = (typeof ORDER_STATUS)[keyof typeof ORDER_STATUS]


export const USER_STATUS = {
  ACTIVE: 'ACTIVE',
  BANNED: 'BANNED',
  SUSPENDED: 'SUSPENDED'
} as const

export type USER_STATUS = (typeof USER_STATUS)[keyof typeof USER_STATUS]


export const PAYMENT_METHOD = {
  CASH_ON_DELIVERY: 'CASH_ON_DELIVERY'
} as const

export type PAYMENT_METHOD = (typeof PAYMENT_METHOD)[keyof typeof PAYMENT_METHOD]